fixes after review. common code moved to getMetaschemaBuilder
