+++ server/src/main/java/com/epam/aidial/core/server/util/CustomApplicationUtils.java
@@ -2,11 +2,9 @@ package com.epam.aidial.core.server.util;
 
 import com.epam.aidial.core.config.Application;
 import com.epam.aidial.core.config.Config;
-import com.epam.aidial.core.metaschemas.MetaSchemaHolder;
 import com.epam.aidial.core.server.ProxyContext;
 import com.epam.aidial.core.server.security.EncryptionService;
 import com.epam.aidial.core.server.validation.CustomAppValidationException;
-import com.epam.aidial.core.server.validation.DialFileFormat;
 import com.epam.aidial.core.server.validation.DialFileKeyword;
 import com.epam.aidial.core.server.validation.DialMetaKeyword;
 import com.epam.aidial.core.server.validation.ListCollector;
@@ -21,7 +19,6 @@ import com.networknt.schema.InputFormat;
 import com.networknt.schema.JsonMetaSchema;
 import com.networknt.schema.JsonSchema;
 import com.networknt.schema.JsonSchemaFactory;
-import com.networknt.schema.NonValidationKeyword;
 import com.networknt.schema.ValidationMessage;
 import lombok.experimental.UtilityClass;
 
@@ -33,21 +30,15 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import static com.epam.aidial.core.metaschemas.MetaSchemaHolder.getMetaschemaBuilder;
+
 
 @UtilityClass
 public class CustomApplicationUtils {
 
-    private static final JsonMetaSchema DIAL_META_SCHEMA = JsonMetaSchema.builder(MetaSchemaHolder.CUSTOM_APPLICATION_META_SCHEMA_ID,
-                    JsonMetaSchema.getV7())
-            .keyword(new NonValidationKeyword("dial:custom-application-type-editor-url"))
-            .keyword(new NonValidationKeyword("dial:custom-application-type-display-name"))
-            .keyword(new NonValidationKeyword("dial:custom-application-type-completion-endpoint"))
-            .keyword(new NonValidationKeyword("dial:property-kind"))
-            .keyword(new NonValidationKeyword("dial:property-order"))
-            .keyword(new NonValidationKeyword("$defs"))
+    private static final JsonMetaSchema DIAL_META_SCHEMA = getMetaschemaBuilder()
             .keyword(new DialMetaKeyword())
             .keyword(new DialFileKeyword())
-            .format(new DialFileFormat())
             .build();
 
     private static final JsonSchemaFactory SCHEMA_FACTORY = JsonSchemaFactory.builder()
